pyproject.toml
[project]
name = "meeting-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic>=2.11.9",
    "pymongo>=4.15.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
